Move scenario testing logic into command
