+++ packages/genesys-web-messaging-tester-cli/__tests__/testScenarioCommand/yamlTestScript/scenarios/scenarioSectionValidated.spec.ts
@@ -1,17 +1,16 @@
-import { Cli, createCli } from '../../src/cli';
 import { readFileSync } from 'fs';
 import { Command } from 'commander';
 import stripAnsi from 'strip-ansi';
+import { createCli } from '../../../../src/createCli';
 
 describe('Scenario Section Validated', () => {
   let capturedOutput: {
     errOut: string[];
   };
 
-  let program: Command;
   let fsReadFileSync: jest.MockedFunction<typeof readFileSync>;
 
-  let cli: Cli;
+  let cli: Command;
 
   beforeEach(() => {
     fsReadFileSync = jest.fn();
@@ -23,52 +22,29 @@ describe('Scenario Section Validated', () => {
       errOut: [],
     };
 
-    program = new Command();
-    program.configureOutput({
-      writeErr: (str) => capturedOutput.errOut.push(str),
-    });
+    const cliCommand = new Command()
+      .exitOverride(() => {
+        throw new Error('CLI Command errored');
+      })
+      .configureOutput({
+        writeErr: (str) => capturedOutput.errOut.push(str),
+      });
 
-    cli = createCli({
-      program,
+    const scenarioTestCommand = new Command()
+      .exitOverride(() => {
+        throw new Error('Scenario Test Command errored');
+      })
+      .configureOutput({
+        writeErr: (str) => capturedOutput.errOut.push(str),
+      });
+
+    cli = createCli(cliCommand, {
+      command: scenarioTestCommand,
       fsReadFileSync,
       fsAccessSync: jest.fn(),
       webMessengerSessionFactory: jest.fn().mockReturnValue(webMessengerSession),
       conversationFactory: jest.fn().mockReturnValue(conversation),
-      processExitOverride: () => {
-        throw new Error('force app to exit');
-      },
     });
-  });
-
-  test('config section is optional', async () => {
-    fsReadFileSync.mockReturnValue(`
-scenarios:
-  scenario:
-    - say: hi from scenario
-`);
-    await expect(
-      cli([
-        ...['node', '/path/to/cli'],
-        ...['--deployment-id', 'test-deployment-id'],
-        ...['--region', 'test-region'],
-        ...['test-path'],
-      ]),
-    ).resolves.toBeUndefined();
-
-    expect(capturedOutput.errOut.map(stripAnsi)).toStrictEqual([]);
-  });
-
-  test('scenarios section is required', async () => {
-    fsReadFileSync.mockReturnValue(`
-config:
-  deploymentId: xx
-  region: xx
-`);
-    await expect(
-      cli([...['node', '/path/to/cli'], ...['test-path']]),
-    ).rejects.toBeDefined();
-
-    expect(capturedOutput.errOut.map(stripAnsi)).toStrictEqual(['"scenarios" is required\n']);
   });
 
   test('scenario step can only contain one element', async () => {
@@ -82,7 +58,7 @@ scenarios:
       waitForReplyContaining: hello
 `);
     await expect(
-      cli([...['node', '/path/to/cli'], ...['test-path']]),
+      cli.parseAsync([...['node', '/path/to/cli'], 'test-scenario', ...['test-path']]),
     ).rejects.toBeDefined();
 
     expect(capturedOutput.errOut.map(stripAnsi)).toStrictEqual([
@@ -100,7 +76,7 @@ scenarios:
     - testing: 123
 `);
     await expect(
-      cli([...['node', '/path/to/cli'], ...['test-path']]),
+      cli.parseAsync([...['node', '/path/to/cli'], 'test-scenario', ...['test-path']]),
     ).rejects.toBeDefined();
 
     expect(capturedOutput.errOut.map(stripAnsi)).toStrictEqual([
