+++ packages/genesys-web-messaging-tester-cli/src/createCli.ts
@@ -1,0 +1,12 @@
+import { Command } from 'commander';
+import {
+  createScenarioTestCommand,
+  ScenarioTestCommandDependencies,
+} from './createScenarioTestCommand';
+
+export function createCli(
+  command: Command = new Command(),
+  scenarioTestCommandDependencies?: ScenarioTestCommandDependencies,
+): Command {
+  return command.addCommand(createScenarioTestCommand(scenarioTestCommandDependencies));
+}
